Update RC-202 File Transfer.py
Simplification (realized files could be written over if they happen to be the same size, which is possible since .wav files are uncompressed and these recordings are in standardized lengths. Complication, computation, & risk not worth it. Review your own wav files before uploading the next batch! Or just put them in a new folder, I'm not your dad.)